[Fix-18269][Master] Move LogicFakeTask to test scope
