+++ dolphinscheduler-master/src/test/java/org/apache/dolphinscheduler/server/master/integration/fake/LogicFakeTaskChannelFactory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.dolphinscheduler.plugin.task.api.task;
+package org.apache.dolphinscheduler.server.master.integration.fake;
 
 import org.apache.dolphinscheduler.plugin.task.api.TaskChannel;
 import org.apache.dolphinscheduler.plugin.task.api.TaskChannelFactory;
