+++ dolphinscheduler-master/src/test/java/org/apache/dolphinscheduler/server/master/integration/fake/LogicFakeTaskParameters.java
@@ -15,19 +15,32 @@
  * limitations under the License.
  */
 
-package org.apache.dolphinscheduler.plugin.task.api.task;
+package org.apache.dolphinscheduler.server.master.integration.fake;
 
-import org.apache.dolphinscheduler.common.utils.JSONUtils;
 import org.apache.dolphinscheduler.plugin.task.api.parameters.AbstractParameters;
-import org.apache.dolphinscheduler.plugin.task.api.parameters.LogicFakeTaskParameters;
+
+import org.apache.commons.lang3.StringUtils;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import com.google.common.annotations.VisibleForTesting;
 
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 @VisibleForTesting
-public class LogicFakeTaskChannel extends AbstractLogicTaskChannel {
+public class LogicFakeTaskParameters extends AbstractParameters {
+
+    private String shellScript;
 
     @Override
-    public AbstractParameters parseParameters(String taskParams) {
-        return JSONUtils.parseObject(taskParams, LogicFakeTaskParameters.class);
+    public boolean checkParameters() {
+        if (StringUtils.isEmpty(shellScript)) {
+            throw new IllegalArgumentException("shellScript is null or empty");
+        }
+        return true;
     }
+
 }
